Refactoring directory structure

+++ core/bin/scaffold-tests.sh
@@ -4,7 +4,7 @@ set -e;
 
 themedir=$(pwd)
 theme=$(basename $themedir)
-if [ ! -e style.css ]; then
+if [ ! -e resources/style.css ]; then
   echo 'Current directory is not a theme.'
   echo $(pwd)
   exit 1
@@ -23,7 +23,8 @@ phar extract -f $(which wp) "$themedir/core/bin/wp.phar">/dev/null 2>&1
 wpclidir="$themedir/core/bin/wp.phar$(which wp)"
 
 if [ ! -e "$themedir/core/bin" ]; then
-	mkdir -p "$themedir/core/bin"
+	echo "$themedir/core/bin is not exsists."
+  exit 1;
 fi
 
 cp -f "$wpclidir/templates/install-wp-tests.sh" "$themedir/core/bin/install-wp-tests.sh"
@@ -32,13 +33,23 @@ cp -f "$wpclidir/templates/phpunit.xml.dist" "$themedir/phpunit.xml"
 
 if [ ! -e "$themedir/tests" ]; then
 	mkdir "$themedir/tests"
+else
+  echo "$themedir/tests is alerady exsists."
 fi
 
-cp -f "$wpclidir/templates/plugin-bootstrap.mustache" "$themedir/tests/plugin-bootstrap.mustache"
-sed -e "s/require dirname( dirname( __FILE__ ) ) \. '\/{{plugin_slug}}\.php';/register_theme_directory( dirname( __FILE__ ) . '\/\.\.\/\.\.\/' ); switch_theme('$theme'); search_theme_directories();/g" "$themedir/tests/plugin-bootstrap.mustache">"$themedir/tests/bootstrap.php"
-rm -f "$themedir/tests/plugin-bootstrap.mustache"
+if [ ! -e "$themedir/tests/bootstrap.php" ]; then
+  cp -f "$wpclidir/templates/plugin-bootstrap.mustache" "$themedir/tests/plugin-bootstrap.mustache"
+  sed -e "s/require dirname( dirname( __FILE__ ) ) \. '\/{{plugin_slug}}\.php';/register_theme_directory( dirname( __FILE__ ) . '\/\.\.\/\.\.\/' ); switch_theme('$theme\/resources'); search_theme_directories();/g" "$themedir/tests/plugin-bootstrap.mustache">"$themedir/tests/bootstrap.php"
+  rm -f "$themedir/tests/plugin-bootstrap.mustache"
+else
+  echo "$themedir/tests/bootstrap.php is alerady exsists."
+fi
 
-cp -f "$wpclidir/templates/plugin-test-sample.mustache" "$themedir/tests/test-sample.php"
+if [ ! -e "$themedir/tests/test-sample.php" ]; then
+  cp -f "$wpclidir/templates/plugin-test-sample.mustache" "$themedir/tests/test-sample.php"
+else
+  echo "$themedir/tests/test-sample.php is alerady exsists."
+fi
 
-rm -rf "$themedir/corebin/wp.phar"
+rm -rf "$themedir/core/bin/wp.phar"
 echo "done!"
